projects/analytics-au/pyproject.toml
[project]
name = "analytics_au"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster-duckdb-pandas>=0.28.14",
    "dagster==1.12.14",
    "requests>=2.32.5",
    "openpyxl>=3.1.5",
    "trendspyg>=0.4.2",
    "dagster-dbt>=0.28.14",
    "dbt-duckdb>=1.10.1",
    "beautifulsoup4>=4.14.3",
    "pdfplumber>=0.11.9",
    "dagster-dlt>=0.28.14",
    "dlt[duckdb]>=1.6.1",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "analytics_au"
registry_modules = [
    "analytics_au.components.*",
]


